requirements.txt
openai-whisper==20231117
sounddevice==0.4.6
numpy==1.24.3
python-docx==1.0.1
PySide6==6.6.1
SpeechRecognition==3.10.0
pyaudio==0.2.13